+++ tests/testthat/test-read_hashed_file.R
@@ -1,6 +1,6 @@
 test_that("read_hashed_file works for file with matching hash", {
   file_path <- "testdata/test_data.parquet"
-  hash <- digest::digest(file = file_path)
+  hash <- digest::digest(file = file_path, algo = "blake3")
   expect_no_condition(read_hashed_file(file_path, hash))
 })
 
@@ -12,20 +12,20 @@ test_that("read_hashed_file fails for mismatching hash", {
 
 test_that("read_hashed_file fails for unsupported file type", {
   file_path <- "testdata/test_data.txt"
-  hash <- digest::digest(file = file_path)
+  hash <- digest::digest(file = file_path, algo = "blake3")
   expect_warning(read_hashed_file(file_path, hash))
 })
 
 test_that("read_hashed_file reads csvs too", {
   file_path <- "testdata/test_data.csv"
-  hash <- digest::digest(file = file_path)
+  hash <- digest::digest(file = file_path, algo = "blake3")
 
   expect_no_error(read_hashed_file(file_path, hash))
 })
 
 test_that("read_hashed_file reads sas file", {
   file_path <- "testdata/test_data.sas7bdat"
-  hash <- digest::digest(file = file_path)
+  hash <- digest::digest(file = file_path, algo = "blake3")
 
   expect_no_error(read_hashed_file(file_path, hash))
 })