+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * Added recycling warning to all functions.
 * Added `ibw` (ideal body weight) and `aibw` (adjusted ideal body weight) functions.
-* Added `obesity_category` function
+* Added `bmic` (BMI category) function
 * Added `agec` (age category) function
 * Added `cor_df` function that takes a dataframe and a vector of column names to compute all pairwise correlations.
 * Updated `egfr` equation default to use CKDEPI 2021 equation.